+++ src/main/java/com/food_management/controllers/UserIngredientController.java
@@ -1,6 +1,8 @@
 package com.food_management.controllers;
 
 import com.food_management.dtos.IngredientInFridgeAndRecipeDto;
+import com.food_management.exceptions.InactiveAccountException;
+import com.food_management.security.UserSessionService;
 import com.food_management.services.impl.UserIngredientServiceImpl;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,6 +18,7 @@ import java.util.List;
 @Transactional
 public class UserIngredientController {
 
+    private UserSessionService userSessionService;
     private UserIngredientServiceImpl service; //TODO: usunac impl
 
     public UserIngredientController(UserIngredientServiceImpl service) {
@@ -25,6 +28,10 @@ public class UserIngredientController {
     @PreAuthorize("hasAnyAuthority('USER')")
     @RequestMapping(method = RequestMethod.GET)
     ResponseEntity<List<IngredientInFridgeAndRecipeDto>> findAll() {
+        if(!userSessionService.isActive()){
+            throw new InactiveAccountException("Inactive account.");
+        }
+
         List<IngredientInFridgeAndRecipeDto> dtoList = service.findAll();
         return ResponseEntity.ok(dtoList);
     }
@@ -32,6 +39,10 @@ public class UserIngredientController {
     @PreAuthorize("hasAnyAuthority('USER')")
     @RequestMapping(method = RequestMethod.POST)
     ResponseEntity add(@Valid @RequestBody IngredientInFridgeAndRecipeDto dto) throws Exception {
+        if(!userSessionService.isActive()){
+            throw new InactiveAccountException("Inactive account.");
+        }
+
         IngredientInFridgeAndRecipeDto created = service.add(dto);
         return new ResponseEntity<IngredientInFridgeAndRecipeDto>(created, HttpStatus.CREATED);
     }
@@ -39,6 +50,10 @@ public class UserIngredientController {
     @PreAuthorize("hasAnyAuthority('USER')")
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     ResponseEntity delete(@PathVariable Long id) {
+        if(!userSessionService.isActive()){
+            throw new InactiveAccountException("Inactive account.");
+        }
+
         service.delete(id);
         return ResponseEntity.ok().build();
     }
@@ -46,6 +61,10 @@ public class UserIngredientController {
     @PreAuthorize("hasAnyAuthority('USER')")
     @RequestMapping(method = RequestMethod.PUT)
     ResponseEntity update(@Valid @RequestBody IngredientInFridgeAndRecipeDto dto) {
+        if(!userSessionService.isActive()){
+            throw new InactiveAccountException("Inactive account.");
+        }
+
         return ResponseEntity.ok(service.update(dto));
     }
